optimized travis script
The new script will generate less travis jobs thus improving build time and reducing traving load.
The downside is that the error messages are little less informative. On the other hand at the
end of the log of each board there will be a list of all successful and failed example builds.